refactor: migrating data fetching directly to the server components
